[Refactor] PointService - PointUseService 와 통합
