+++ src/main/java/com/inq/wishhair/wesharewishhair/user/service/PointService.java
@@ -5,6 +5,7 @@ import com.inq.wishhair.wesharewishhair.user.domain.User;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointHistory;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointRepository;
 import com.inq.wishhair.wesharewishhair.user.domain.point.PointType;
+import com.inq.wishhair.wesharewishhair.user.service.dto.MailDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,10 +15,29 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class PointService {
 
+    private static final String MAIL_TITLE = "We-Share-Wish-Hair 포인트 환급 요청";
+
     private final PointRepository pointRepository;
+    private final UserFindService userFindService;
+    private final MailSendService mailSendService;
 
     @Transactional
-    public void insertPointHistory(PointType pointType, int dealAmount, User user) {
+    public void usePoint(PointUseRequest request, Long userId) {
+
+        User user = userFindService.findByUserId(userId);
+        insertPointHistory(PointType.USE, request.getDealAmount(), user);
+
+        sendRefundRequestMail(request, user);
+    }
+
+    @Transactional
+    public void chargePoint(int dealAmount, Long userId) {
+
+        User user = userFindService.findByUserId(userId);
+        insertPointHistory(PointType.CHARGE, dealAmount, user);
+    }
+
+    private void insertPointHistory(PointType pointType, int dealAmount, User user) {
 
         user.updateAvailablePoint(pointType, dealAmount);
         PointHistory pointHistory = generatePointHistory(pointType, dealAmount, user);
@@ -33,6 +53,20 @@ public class PointService {
             point = user.getAvailablePoint() - dealAmount;
         }
         return PointHistory.createPointHistory(user, pointType, dealAmount, point);
+    }
+
+    private void sendRefundRequestMail(PointUseRequest request, User user) {
+        String contents = generateContents(request, user);
+        MailDto mailDto = new MailDto(user.getEmail(), MAIL_TITLE, contents);
+
+        mailSendService.sendMail(mailDto);
+    }
+
+    private String generateContents(PointUseRequest request, User user) {
+        String contents = "사용자 이름 : " + user.getName() + "\n";
+        contents += "계좌 번호 : " + request.getAccountNumber() + "\n";
+        contents += "환급 포인트 : " + request.getDealAmount();
+        return contents;
     }
 }
